linux-2.6.0

+++ drivers/media/video/bw-qcam.c
@@ -72,7 +72,6 @@ OTHER DEALINGS IN THE SOFTWARE.
 #include <linux/mm.h>
 #include <linux/parport.h>
 #include <linux/sched.h>
-#include <linux/version.h>
 #include <linux/videodev.h>
 #include <asm/semaphore.h>
 #include <asm/uaccess.h>
@@ -856,8 +855,8 @@ static int qcam_ioctl(struct inode *inode, struct file *file,
 	return video_usercopy(inode, file, cmd, arg, qcam_do_ioctl);
 }
 
-static int qcam_read(struct file *file, char *buf,
-		     size_t count, loff_t *ppos)
+static ssize_t qcam_read(struct file *file, char *buf,
+			 size_t count, loff_t *ppos)
 {
 	struct video_device *v = video_devdata(file);
 	struct qcam_device *qcam=(struct qcam_device *)v;
